Factor things.
